pyproject.toml
[project]
name = "meme-generator-agent"
version = "0.1.0"
description = "AI Meme Generator Agent that uses browser automation to create memes based on text prompts."
authors = [{ name = "Paras Chamoli", email = "paras25619@gmail.com" }]
readme = "README.md"
keywords = ['python', 'agent', 'browser-use', 'meme']
requires-python = ">=3.12,<3.13"
dependencies = [
    "bindu==2026.9.2.2",
    "browser-use>=0.1.0",
    "langchain-openai>=0.1.0",
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "pre-commit",
]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://github.com/Paraschamoli/meme-generator-agent"
Repository = "https://github.com/Paraschamoli/meme-generator-agent"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "ty>=0.0.18",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["meme_generator_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 120
fix = true

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
